+++ cockpit_api/src/main/java/com/cockpit/api/controller/SprintController.java
@@ -1,18 +1,11 @@
 package com.cockpit.api.controller;
 
+import com.cockpit.api.service.AuthService;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.CrossOrigin;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 import com.cockpit.api.exception.ResourceNotFoundException;
 import com.cockpit.api.model.dao.Jira;
@@ -27,35 +20,47 @@ import com.cockpit.api.service.SprintService;
 public class SprintController {
     private final SprintService sprintService;
     private final JiraService jiraService;
-	private ModelMapper modelMapper = new ModelMapper();
+    private final AuthService authService;
+    private ModelMapper modelMapper = new ModelMapper();
 	
     @Autowired
     public SprintController(
-            SprintService sprintService,  JiraService jiraService
+            SprintService sprintService,  JiraService jiraService, AuthService authService
     ) {
         this.sprintService = sprintService;
-        this.jiraService=jiraService;
+        this.jiraService = jiraService;
+        this.authService = authService;
     }
 
     // CREATE a new Sprint
     @PostMapping(
             value = "/api/v1/sprint/create"
     )
-    public ResponseEntity<SprintDTO> createSprint(@RequestBody SprintDTO sprintDTO) {
-        SprintDTO newSprint = sprintService.createNewSprint(sprintDTO);
-        return ResponseEntity.ok(newSprint);
+    public ResponseEntity createSprint(@RequestBody SprintDTO sprintDTO,
+                                                  @RequestHeader("Authorization") String authHeader) {
+        if (authService.isUserAuthorized(authHeader)) {
+            SprintDTO newSprint = sprintService.createNewSprint(sprintDTO);
+            return ResponseEntity.ok(newSprint);
+        } else {
+            return ResponseEntity.ok(HttpStatus.UNAUTHORIZED);
+        }
     }
 
     // GET Sprint BY ID
     @GetMapping(
             value = "/api/v1/sprint/{id}"
     )
-    public ResponseEntity getSprintById(@PathVariable Long id) {
-        try {
-            SprintDTO sprintFound = sprintService.findSprintById(id);
-            return ResponseEntity.ok().body(sprintFound);
-        } catch (com.cockpit.api.exception.ResourceNotFoundException e) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(e.getMessage());
+    public ResponseEntity getSprintById(@PathVariable Long id,
+                                        @RequestHeader("Authorization") String authHeader) {
+        if (authService.isUserAuthorized(authHeader)) {
+            try {
+                SprintDTO sprintFound = sprintService.findSprintById(id);
+                return ResponseEntity.ok().body(sprintFound);
+            } catch (com.cockpit.api.exception.ResourceNotFoundException e) {
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body(e.getMessage());
+            }
+        } else {
+            return ResponseEntity.ok(HttpStatus.UNAUTHORIZED);
         }
     }
 
@@ -63,12 +68,18 @@ public class SprintController {
     @PutMapping(
             value = "/api/v1/sprint/update/{id}"
     )
-    public ResponseEntity updateSprint(@RequestBody SprintDTO sprintDTO, @RequestParam Long id) {
-        try {
-            SprintDTO sprintUpdated = sprintService.updateSprint(sprintDTO, id);
-            return ResponseEntity.ok().body(sprintUpdated);
-        } catch (com.cockpit.api.exception.ResourceNotFoundException e) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(e.getMessage());
+    public ResponseEntity updateSprint(@RequestBody SprintDTO sprintDTO,
+                                       @RequestParam Long id,
+                                       @RequestHeader("Authorization") String authHeader) {
+        if (authService.isUserAuthorized(authHeader)) {
+            try {
+                SprintDTO sprintUpdated = sprintService.updateSprint(sprintDTO, id);
+                return ResponseEntity.ok().body(sprintUpdated);
+            } catch (com.cockpit.api.exception.ResourceNotFoundException e) {
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body(e.getMessage());
+            }
+        } else {
+            return ResponseEntity.ok(HttpStatus.UNAUTHORIZED);
         }
     }
 
@@ -76,25 +87,36 @@ public class SprintController {
     @DeleteMapping(
             value = "/api/v1/sprint/delete/{id}"
     )
-    public ResponseEntity<String> deleteSprint(@PathVariable Long id) {
-        try {
-            sprintService.deleteSprint(id);
-            return ResponseEntity.ok("One Sprint has been deleted");
-        } catch (ResourceNotFoundException e) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(e.getMessage());
+    public ResponseEntity deleteSprint(@PathVariable Long id,
+                                               @RequestHeader("Authorization") String authHeader) {
+        if (authService.isUserAuthorized(authHeader)) {
+            try {
+                sprintService.deleteSprint(id);
+                return ResponseEntity.ok("One Sprint has been deleted");
+            } catch (ResourceNotFoundException e) {
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body(e.getMessage());
+            }
+        } else {
+            return ResponseEntity.ok(HttpStatus.UNAUTHORIZED);
         }
     }
     // GET Sprint BY Sprint Number And Jira
     @GetMapping(
             value = "/api/v1/sprint/{jiraId}/{sprintNumber}"
     )
-    public ResponseEntity getSprintBySprintNumberAndJira(@PathVariable Long jiraId, @PathVariable int sprintNumber) {
-        try {
-        	JiraDTO jira = jiraService.findJiraById(jiraId);
-            Sprint sprintFound = sprintService.findByMvpAndSprintNumber(modelMapper.map(jira, Jira.class), sprintNumber);
-            return ResponseEntity.ok().body(sprintFound);
-        } catch (com.cockpit.api.exception.ResourceNotFoundException e) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(e.getMessage());
+    public ResponseEntity getSprintBySprintNumberAndJira(@PathVariable Long jiraId,
+                                                         @PathVariable int sprintNumber,
+                                                         @RequestHeader("Authorization") String authHeader) {
+        if (authService.isUserAuthorized(authHeader)) {
+            try {
+                JiraDTO jira = jiraService.findJiraById(jiraId);
+                Sprint sprintFound = sprintService.findByMvpAndSprintNumber(modelMapper.map(jira, Jira.class), sprintNumber);
+                return ResponseEntity.ok().body(sprintFound);
+            } catch (com.cockpit.api.exception.ResourceNotFoundException e) {
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body(e.getMessage());
+            }
+        } else {
+            return ResponseEntity.ok(HttpStatus.UNAUTHORIZED);
         }
     }
 }
